refactor(components): [color-picker] switch to script-setup syntax (#9672)
