+++ packages/components/color-picker/__tests__/color-picker.test.tsx
@@ -2,7 +2,7 @@ import { nextTick, ref } from 'vue'
 import { mount } from '@vue/test-utils'
 import { describe, expect, it, vi } from 'vitest'
 import { ElFormItem } from '@element-plus/components/form'
-import ColorPicker from '../src/index.vue'
+import ColorPicker from '../src/color-picker.vue'
 import type { ComponentPublicInstance } from 'vue'
 
 vi.mock('lodash-unified', async () => {
