+++ packages/components/color-picker/index.ts
@@ -1,13 +1,8 @@
-import ColorPicker from './src/index.vue'
+import { withInstall } from '@element-plus/utils'
 
-import type { App } from 'vue'
-import type { SFCWithInstall } from '@element-plus/utils'
+import ColorPicker from './src/color-picker.vue'
 
-ColorPicker.install = (app: App): void => {
-  app.component(ColorPicker.name, ColorPicker)
-}
+export const ElColorPicker = withInstall(ColorPicker)
+export default ElColorPicker
 
-const _ColorPicker = ColorPicker as SFCWithInstall<typeof ColorPicker>
-
-export default _ColorPicker
-export const ElColorPicker = _ColorPicker
+export * from './src/color-picker'
